refactroing one dimension(=interval element)
